Customer manages Orders

+++ app/Http/Controllers/OrderItemController.php
@@ -7,7 +7,7 @@ use App\Http\Requests\StoreOrderItemRequest;
 use App\Http\Requests\UpdateOrderItemRequest;
 use App\Models\Order;
 use App\Models\Product;
-use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Http\Request;
 
 class OrderItemController extends Controller
 {
@@ -30,49 +30,48 @@ class OrderItemController extends Controller
         }
         return response()->json($payload, $code);
     }
-
-
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \App\Http\Requests\StoreOrderItemRequest  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(StoreOrderItemRequest $request)
-    {
-        $product = Product::findOrFail($request->product_id);
-        $order = Order::findOrFail($request->order_id);
-        $order->orderItems()->create([
-            'product_id' => $product->id,
-            'quantity' => $request->quantity,
-            'amount' => $product->price * $request->quantity,
-        ]);
-        $order->amount += $product->price * $request->quantity;
-        $order->save();
-        return response()->json($order, 201);
-    }
-
     /**
      * Display the specified resource.
      *
      * @param  \App\Models\OrderItem  $orderItem
      * @return \Illuminate\Http\Response
      */
-    public function show(FormRequest $request)
+    public function show(int $order_id, int $orderItem_id)
     {
         $code = 404;
-        /* $message = "Order items not found";
+        $message = "Order items not found";
         $payload = ['message' => $message];
-        $order = Order::findOrFail($request->order_id);
-        $orderItem = $order->orderItems->find($request->orderItem_id);
+        $order = Order::findOrFail($order_id);
+        $orderItem = $order->orderItems->find($orderItem_id);
 
         if ($orderItem) {
             $payload = ['data' => $orderItem];
             $code = 200;
-        } */
-        return response()->json("joder", $code);
+        }
+        return response()->json($payload, $code);
     }
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \App\Http\Requests\StoreOrderItemRequest  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(int $order_id)
+    {
+        $quantity = request('quantity');
+        $order = Order::findOrFail($order_id);
+        $product = Product::findOrFail(request('product_id'));
+        $order->orderItems()->create([
+            'product_id' => $product->id,
+            'quantity' => $quantity,
+            'amount' => $product->price * $quantity,
+        ]);
+        $order->amount += $product->price * $quantity;
+        $order->save();
+        return response()->json($order->orderItems, 201);
+    }
+
+
 
     /**
      * Update the specified resource in storage.
@@ -92,13 +91,21 @@ class OrderItemController extends Controller
      * @param  \App\Models\OrderItem  $orderItem
      * @return \Illuminate\Http\Response
      */
-    public function destroy(int $id)
+    public function destroy(int $order_id, int $orderItem_id)
     {
-        $orderItem = OrderItem::findOrFail($id);
-        $order = Order::findOrFail($orderItem->order_id);
-        $order->amount -= $orderItem->amount;
-        $order->save();
-        $orderItem->delete();
-        return response()->json($order, 200);
+        $code = 404;
+        $messageNotFound = "Order items not found";
+        $messageFounded = "Order item deleted";
+        $payload = ['message' => $messageNotFound];
+        $order = Order::findOrFail($order_id);
+        $orderItem = $order->orderItems->find($orderItem_id);
+
+        if ($orderItem) {
+            $order->amount -= $orderItem->amount;
+            $order->save();
+            $orderItem->delete();
+            $payload = ['message' => $messageFounded];
+        }
+        return response()->json($payload, $code);
     }
 }
